baselines/HybridCBM/requirements.txt
apricot_select==0.6.1
click==8.1.7
ftfy==6.3.1
geomloss==0.2.6
lightning==2.4.0
matplotlib==3.9.2
mmengine==0.10.5
mpire==2.10.2
nltk==3.9.1
numpy==2.1.3
openai==1.54.4
packaging==24.2
pandas==2.2.3
Pillow==11.0.0
regex==2024.9.11
scikit_learn==1.5.2
scipy==1.14.1
torch==2.4.1
torchmetrics==1.4.2
torchvision==0.19.1
tqdm==4.66.5
transformers==4.44.2
